use first-class polymorphism for request handlers
Record fields and object methods can have a polymorphic type of the form

    'a . 'a request -> 'a response

which is precisely required to correctly type-check the parameter of
the 'listen' function. This is a polymorphic type (it means: "forall
'a, ..."), so the instantiation of the parameter ('a) is local to the
callsite, and thus it does not let the existential parameter of the
request type escape.